Split Resource tests into happy-path + FailureTest (HYDRAN-2367)
