+++ src/test/java/se/sundsvall/postportalservice/api/AttachmentResourceTest.java
@@ -11,16 +11,13 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
-import se.sundsvall.dept44.problem.violations.ConstraintViolationProblem;
 import se.sundsvall.postportalservice.Application;
 import se.sundsvall.postportalservice.service.AttachmentService;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
-import static se.sundsvall.postportalservice.TestDataFactory.INVALID_MUNICIPALITY_ID;
 import static se.sundsvall.postportalservice.TestDataFactory.MUNICIPALITY_ID;
 
 @SpringBootTest(classes = Application.class, webEnvironment = RANDOM_PORT)
@@ -63,26 +60,5 @@ class AttachmentResourceTest {
 			.expectBody(byte[].class).isEqualTo(content);
 
 		verify(attachmentServiceMock).getAttachmentData(MUNICIPALITY_ID, attachmentId);
-	}
-
-	@Test
-	void getAttachmentsByMessageId_BadRequest() {
-		final var response = webTestClient.get()
-			.uri(uriBuilder -> uriBuilder.replacePath("/{municipalityId}/attachments/{attachmentId}")
-				.build(INVALID_MUNICIPALITY_ID, "invalid"))
-			.exchange()
-			.expectStatus().isBadRequest()
-			.expectBody(ConstraintViolationProblem.class)
-			.returnResult()
-			.getResponseBody();
-
-		assertThat(response.getTitle()).isEqualTo("Constraint Violation");
-		assertThat(response.getViolations()).satisfiesExactlyInAnyOrder(violation -> {
-			assertThat(violation.field()).isEqualTo("downloadAttachment.municipalityId");
-			assertThat(violation.message()).isEqualTo("not a valid municipality ID");
-		}, violation -> {
-			assertThat(violation.field()).isEqualTo("downloadAttachment.attachmentId");
-			assertThat(violation.message()).isEqualTo("not a valid UUID");
-		});
 	}
 }
